Move analysis commands under analyze category
