+++ tests/cli/help.test.ts
@@ -17,12 +17,12 @@ describe("CLI help", () => {
   it("shows concrete examples in sample and eval help", () => {
     const program = createProgram();
     const sampleHelp = program.commands.find((command) => command.name() === "sample")?.helpInformation();
-    const evalHelp = program.commands.find((command) => command.name() === "eval")?.helpInformation();
+    const analyzeCmd = program.commands.find((command) => command.name() === "analyze");
+    const evalHelp = analyzeCmd?.commands.find((command) => command.name() === "eval")?.helpInformation();
 
     expect(sampleHelp).toContain("Select tests without executing");
     expect(sampleHelp).toContain("flaker sample");
     expect(sampleHelp).toContain("Strategies");
     expect(evalHelp).toContain("Measure whether local sampled runs predict CI");
-    expect(evalHelp).toContain("flaker eval --json");
   });
 });
