+++ tests/commands/flaky-variant.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { DuckDBStore } from "../../src/cli/storage/duckdb.js";
-import { runFlakyByVariant } from "../../src/cli/commands/flaky.js";
+import { runFlakyByVariant } from "../../src/cli/commands/analyze/flaky.js";
 
 describe("flaky by variant", () => {
   let store: DuckDBStore;
