+++ tests/integration-e2e-flow.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { DuckDBStore } from "../src/cli/storage/duckdb.js";
 import { runSample } from "../src/cli/commands/sample.js";
-import { runEval } from "../src/cli/commands/eval.js";
+import { runEval } from "../src/cli/commands/analyze/eval.js";
 import { SimpleResolver } from "../src/cli/resolvers/simple.js";
 import type { WorkflowRun, TestResult } from "../src/cli/storage/types.js";
