+++ tests/pipeline/accumulation.test.ts
@@ -5,7 +5,7 @@ import { loadFixtureIntoStore } from "../../src/cli/eval/fixture-loader.js";
 import { analyzeProject, recommendSampling } from "../../src/cli/commands/collect/calibrate.js";
 import { trainModel } from "../../src/cli/commands/train.js";
 import { planSample } from "../../src/cli/commands/sample.js";
-import { runInsights } from "../../src/cli/commands/insights.js";
+import { runInsights } from "../../src/cli/commands/analyze/insights.js";
 import { mkdirSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
